Modified test cases

+++ app/test/utils/test_utils.py
@@ -1,13 +1,14 @@
 import json
-from app.test.setup_test_auth import SetupTestDB
+from app.test.setup_test_db import SetupTestDB
 
 
 class TestCase(SetupTestDB):
 
     def tearDown(self):
         active_col = self.test_db.list_collection_names()
-        for collection in active_col:
-            self.test_db[collection].drop()
+        if len(active_col) > 0:
+            for collection in active_col:
+                self.test_db[collection].drop()
 
     def test_welcome_endpoint(self):
         res = self.app_client.get('/')
@@ -16,61 +17,60 @@ class TestCase(SetupTestDB):
         self.assertTrue(data == 'The Hunters Company')
 
     def test_post_roles(self):
-        role_data = {'type': 'test1', 'description': 'test case'}
-        res = self.app_client.post('/roles', data=role_data)
+        payload = {'type': 'test1', 'description': 'test case'}
+        res = self.app_client.post('/roles', data=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['result']['type'], 'test1')
 
-        role_data = {'type': 'test2', 'description': 'test case'}
-        res = self.app_client.post('/roles', json=role_data)
+        payload = {'type': 'test2', 'description': 'test case'}
+        res = self.app_client.post('/roles', json=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['result']['type'], 'test2')
 
     def test_fail_duplicate_post_roles(self):
-        role_data = {'type': 'test1', 'description': 'test case'}
-        res = self.app_client.post('/roles', data=role_data)
+        payload = {'type': 'test1', 'description': 'test case'}
+        res = self.app_client.post('/roles', data=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['result']['type'], 'test1')
 
-        res = self.app_client.post('/roles', json=role_data)
+        res = self.app_client.post('/roles', json=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
         self.assertEqual(data['result'], 'unprocessable')
 
     def test_post_client_type(self):
-        request_data = {'employment_type': 'test1'}
-        res = self.app_client.post('/client_types', data=request_data)
+        payload = {'employment_type': 'test1'}
+        res = self.app_client.post('/client_types', data=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['result']['employment_type'], 'test1')
 
-        request_data = {'employment_type': 'test2'}
-        res = self.app_client.post('/client_types', json=request_data)
+        payload = {'employment_type': 'test2'}
+        res = self.app_client.post('/client_types', json=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['result']['employment_type'], 'test2')
 
     def test_fail_duplicate_post_client_type(self):
-        request_data = {'employment_type': 'test1'}
-        res = self.app_client.post('/client_types', data=request_data)
+        payload = {'employment_type': 'test1'}
+        res = self.app_client.post('/client_types', data=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['result']['employment_type'], 'test1')
 
-        request_data = {'employment_type': 'test1'}
-        res = self.app_client.post('/client_types', json=request_data)
+        payload = {'employment_type': 'test1'}
+        res = self.app_client.post('/client_types', json=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
         self.assertTrue(data['result'], 'unprocessable')
-
